+++ packages/nucleo/src/importacion/index.ts
@@ -1,2 +1,3 @@
 export * from './csv'
 export * from './miclase'
+export * from './xlsx'